ENH: BCut Test added for OpenCV updates
This set of changes allow for testing
different versions of OpenCV to ensure
that updates do not cause regressions
in the the results.
