+++ BRAINSCut/TestSuite/NetConfigurations/OpenCVStabilityTest_caudate.xml.in
@@ -1,0 +1,122 @@
+<!-- ==========================================================================
+
+  - BRAINSCut Test XML File
+  - Author : Eun Young( Regina ) Kim
+  - 2014 Apr
+
+    /scratch/eunyokim/src/NamicExternal201404/build_release/bin
+    ========================================================================== -->
+
+<!-- This is the part holding everything -->
+<AutoSegProcessDescription>
+
+  <!-- ==========================================================================
+        Registration Description
+       ========================================================================== -->
+  <RegistrationConfiguration
+    ImageTypeToUse="T1"
+    ID="BSpline"
+    BRAINSROIAutoDilateSize="10"
+    />
+  <!-- ==========================================================================
+        Template Data Scan Description
+       ========================================================================== -->
+  <DataSet Name="template"  Type="Atlas" >
+    <Image Type="T1" Filename="@Reference_BINARY_DIR@/template_t1.nii.gz"/>
+    <Image Type="T2" Filename="na" />
+    <Image Type="SG" Filename="na" />
+    <Mask Type="RegistrationROI" Filename="@Reference_BINARY_DIR@/template_t1.nii.gz" />
+    <SpatialLocation Type="rho"   Filename="@Reference_BINARY_DIR@/spatialImages/rho.nii.gz" />
+    <SpatialLocation Type="phi"   Filename="@Reference_BINARY_DIR@/spatialImages/phi.nii.gz" />
+    <SpatialLocation Type="theta" Filename="@Reference_BINARY_DIR@/spatialImages/theta.nii.gz" />
+  </DataSet>
+
+  <!-- ==========================================================================
+        Probability Map  Description
+       ========================================================================== -->
+  <ProbabilityMap
+         StructureID="l_caudate"
+         Gaussian="1.0"
+         GenerateVector="true"
+         Filename="@Reference_BINARY_DIR@/probabilityMaps/l_caudate_ProbabilityMap.nii.gz"
+
+  />
+  <ProbabilityMap
+         StructureID="r_caudate"
+         Gaussian="1.0"
+         GenerateVector="true"
+         Filename="@Reference_BINARY_DIR@/probabilityMaps/r_caudate_ProbabilityMap.nii.gz"
+
+  />
+
+  <!-- ==========================================================================
+        ANN Description
+       ========================================================================== -->
+  <ANNParameters Iterations             = "10"
+                    MaximumVectorsPerEpoch = "10000"
+                    EpochIterations        = "100"
+                    ErrorInterval          = "1"
+                    DesiredError           = "0.000001"
+                    NumberOfHiddenNodes    = "60"
+                    ActivationSlope        = "1.0"
+                    ActivationMinMax       = "1.0"
+   />
+
+  <!-- ==========================================================================
+        ANN Description DUMMY
+       ========================================================================== -->
+   <NeuralNetParams MaskSmoothingValue     = "0.0"
+                    GradientProfileSize    = "1"
+                    TrainingVectorFilename = "/scratch/eunyokim/src/NamicExternal201404/build_release/BRAINSTools-build/BRAINSCut/TestSuite/trainVectorHE.txt"
+                    TrainingModelFilename  = "/scratch/eunyokim/src/NamicExternal201404/build_release/BRAINSTools-build/BRAINSCut/TestSuite/trainModel.txt"
+                    TestVectorFilename     = "na"
+                    Normalization          = "Linear"
+   />
+
+  <!-- ==========================================================================
+        ANN Description
+       ========================================================================== -->
+    <ApplyModel     MaskThresh             = "0.01"
+                    GaussianSmoothingSigma = "3"
+    />
+
+  <!-- ==========================================================================
+        Random Forest Description
+       ========================================================================== -->
+    <RandomForestParameters MaxDepth        = "5"
+                            MinSampleCount    = "5"
+                            UseSurrogates     = "false"
+                            CalcVarImportance = "false"
+                            MaxTreeCount      = "5"
+    />
+
+  <!-- ==========================================================================
+        Apply Data Scan Description
+       ========================================================================== -->
+    <DataSet Name="Scan3"  Type="Apply"
+      OutputDir="@CMAKE_CURRENT_BINARY_DIR@/">
+    <Image Type="T1" Filename="@scan3T1@"/>
+    <Image Type="T2" Filename="@scan3T2@"/>
+    <Image Type="SG" Filename="/raid0/homes/eunyokim/Desktop/TEMP/scan3SG.nii.gz"/>
+    <Mask Type="RegistrationROI" Filename="@scan3T1@" />
+    <Mask Type="l_caudate"  Filename="@CMAKE_CURRENT_BINARY_DIR@/l_caudate_scan3.nii.gz"/>
+    <Mask Type="r_caudate"  Filename="@CMAKE_CURRENT_BINARY_DIR@/r_caudate_scan3.nii.gz"/>
+    <Registration
+      SubjToAtlasRegistrationFilename="@CMAKE_CURRENT_BINARY_DIR@/Scan3ToTemplate.h5"
+      AtlasToSubjRegistrationFilename="@CMAKE_CURRENT_BINARY_DIR@/TemplateToScan3.h5"
+      ID="BSpline"
+      />
+  </DataSet>
+  <!--DataSet Name="Scan4"  Type="Apply"
+    OutputDir="/scratch/eunyokim/src/NamicExternal201404/build_release/BRAINSTools-build/BRAINSCut/TestSuite/MaskOutputDir/Scan4/">
+    <Image Type="T1" Filename="@TestData_DIR@/scan4T1.nii.gz"/>
+    <Image Type="T2" Filename="@TestData_DIR@/scan4T2.nii.gz"/>
+    <Mask Type="RegistrationROI" Filename="@TestData_DIR@/scan4T1.nii.gz" />
+    <Registration
+      SubjToAtlasRegistrationFilename=""
+      AtlasToSubjRegistrationFilename="@TestData_DIR@/Transforms_h5/AtlasToSubjectScan4.h5"
+      ID="BSpline"
+      />
+  </DataSet-->
+
+</AutoSegProcessDescription>
